+++ frontend/screens/account/FindUser.js
@@ -1,6 +1,5 @@
 import React, { useState, useContext } from "react";
 import {
-  signInWithEmailAndPassword,
   createUserWithEmailAndPassword,
 } from "firebase/auth";
 import { auth } from "../../services/firebaseConfig";
@@ -24,71 +23,6 @@ import { AuthenticationContext } from "../../context/AuthenticationContext";
 import globalStyles from "../../assets/styles/Styles.js";
 import Toast from "react-native-toast-message";
 import { saveUserToFirestore, deleteUserDataFromFirestore  } from "../../services/firebaseController.js";
-
-export const UserLogin = ({ isVisible, toggleVisible }) => {
-  const [usermail, setLoginUsername] = useState("");
-  const [password, setLoginPassword] = useState("");
-  const { setAuthState } = useContext(AuthenticationContext);
-  const navigation = useNavigation();
-
-  const handleLogin = async () => {
-    try {
-      const data = await signInWithEmailAndPassword(auth, usermail, password);
-      if (data) {
-        setAuthState(data);
-      } else {
-        Alert.alert(
-          "Kirjautuminen epäonnistui",
-          "Virheellinen käyttäjätunnus tai salasana"
-        );
-      }
-    } catch (error) {
-      console.error("Login error:", error);
-      Alert.alert("Virhe kirjautumisessa", error.message || "Yhteysvirhe");
-      Toast.show({
-        type: "error",
-        text1: "Kirjautuminen epäonnistui",
-        text2: error.message || "Yhteysvirhe",
-      });
-    }
-  };
-
-  return (
-    <View>
-      <Heading
-        title="Kirjaudu sisään"
-        onPress={toggleVisible}
-        style={{
-          backgroundColor: isVisible
-            ? "rgba(25, 26, 30, 0.7)"
-            : "rgba(18, 18, 18, 0.9)",
-        }}
-      />
-      {isVisible && (
-        <AccountSection>
-          <CommonText
-            value="Kirjaudu sisään omalla käyttäjätunnuksellasi ja salasanallasi."
-            editable={false}
-          />
-          <CommonText
-            value={usermail}
-            onChangeText={setLoginUsername}
-            editable={true}
-            trailingIcon={() => <Icon name="person" />}
-          />
-          <CommonText
-            value={password}
-            onChangeText={setLoginPassword}
-            editable={true}
-            trailingIcon={() => <Icon name="lock" />}
-            secureTextEntry
-          />
-          <ButtonContinue title="Kirjaudu" onPress={handleLogin} />
-        </AccountSection>
-      )}
-    </View>
-  );
-};
 
 export const UserRegister = ({ isVisible, toggleVisible }) => {
   const [registerUsername, setRegisterUsername] = useState("");
